🔨 refactor: 인증 시스템 아키텍처 개선 및 불필요한 코드 제거
- api proxy 레이어 제거 및 직접 백엔드 통신 구조로 전환
- header 컴포넌트 서버/클라이언트 분리로 ssr 깜빡임 해결
- 사용하지 않는 스키마, 타입, 유틸리티 함수 정리
- 토큰 갱신 무한 루프 방지 로직 추가
- 세션 에러 자동 로그아웃 처리 구현
