+++ apps/web/app/(public)/home/constants.ts
@@ -1,0 +1,11 @@
+const DURATION = 1.2 as const;
+
+const LINES = [
+  "The single workspace where",
+  "developers ship,",
+  "planners steer, and",
+  "designers shine,",
+  "as one.",
+] as const;
+
+export { DURATION, LINES };
