+++ apps/web/app/(public)/signup/page.tsx
@@ -17,16 +17,14 @@ import {
   Input,
 } from "@triad/ui";
 import Link from "next/link";
-import { useRouter } from "next/navigation";
 import { useForm } from "react-hook-form";
-import { ZodError } from "zod";
 
+import { useAuth } from "@/hooks/use-auth";
 import { SignupRequest, SignupRequestSchema } from "@/schemas/auth";
-import { API_ENDPOINTS, apiClient } from "@/utils/api-client";
-import { formatZodError } from "@/utils/zod-helpers";
 
-export default function RegisterPreview() {
-  const router = useRouter();
+export default function Register() {
+  const { signup } = useAuth();
+
   const form = useForm<SignupRequest>({
     resolver: zodResolver(SignupRequestSchema),
     defaultValues: {
@@ -39,23 +37,7 @@ export default function RegisterPreview() {
   });
 
   async function onSubmit(values: SignupRequest) {
-    try {
-      const validatedValues = SignupRequestSchema.parse(values);
-
-      await apiClient.post(API_ENDPOINTS.AUTH.REGISTER, {
-        json: validatedValues,
-      });
-
-      router.push("/login");
-    } catch (error) {
-      if (error instanceof ZodError) {
-        const formattedError = formatZodError(error);
-        form.setError("root.message", { message: formattedError });
-        return;
-      }
-
-      throw error;
-    }
+    await signup(values);
   }
 
   return (
