+++ apps/web/utils/api-client/index.ts
@@ -143,6 +143,3 @@ export const apiClient = createApiClient();
 
 // API 엔드포인트 상수 재내보내기
 export { API_ENDPOINTS } from "../../constants/api-endpoints";
-
-// 서버 컴포넌트용 API 클라이언트 함수들 재내보내기
-export { createServerApiClient, createServerApiWithAuth } from "./server";
